--- featuretools/primitives/standard/aggregation_primitives.py
@@ -1,110 +1,0 @@
-from datetime import datetime, timedelta
-
-import numpy as np
-import pandas as pd
-from dask import dataframe as dd
-from scipy import stats
-from woodwork.column_schema import ColumnSchema
-from woodwork.logical_types import (
-    Boolean,
-    BooleanNullable,
-    Datetime,
-    Double,
-    IntegerNullable,
-)
-
-from featuretools.primitives.base.aggregation_primitive_base import AggregationPrimitive
-from featuretools.utils import convert_time_units
-from featuretools.utils.gen_utils import Library
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-def convert_datetime_to_floats(x):
-    first = int(x.iloc[0].value * 1e-9)
-    x = pd.to_numeric(x).astype(np.float64).values
-    dividend = find_dividend_by_unit(first)
-    x *= 1e-9 / dividend
-    return x
-
-
-def convert_timedelta_to_floats(x):
-    first = int(x.iloc[0].total_seconds())
-    dividend = find_dividend_by_unit(first)
-    x = pd.TimedeltaIndex(x).total_seconds().astype(np.float64) / dividend
-    return x
-
-
-def find_dividend_by_unit(time):
-    """Finds whether time best corresponds to a value in
-    days, hours, minutes, or seconds.
-    """
-    for dividend in [86400, 3600, 60]:
-        div = time / dividend
-        if round(div) == div:
-            return dividend
-    return 1
-
-
-
